server/artwork_requirements.txt
fastapi
uvicorn[standard]
gunicorn
requests
anthropic
python-dotenv
pydantic
